Move proxy-related classes to their own package
