+++ indicator/src/main/java/com/fuzz/indicator/ViewPagerStateProxy.java
@@ -22,6 +22,11 @@ import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.view.View;
 
+import com.fuzz.indicator.proxy.IndicatorOffsetEvent;
+import com.fuzz.indicator.proxy.ProxyReference;
+import com.fuzz.indicator.proxy.StateProxy;
+import com.fuzz.indicator.proxy.ViewProvidingStateProxy;
+
 /**
  * {@link StateProxy} wrapper around a {@link ViewPager}. All calls
  * are delegated to this object.
