+++ indicator/src/main/java/com/fuzz/indicator/proxy/OffsetEvent.java
@@ -13,12 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fuzz.indicator;
+package com.fuzz.indicator.proxy;
 
 import android.graphics.Matrix;
 import android.text.Spannable;
 import android.widget.ImageView;
 
+import com.fuzz.indicator.CutoutImageCell;
+import com.fuzz.indicator.CutoutViewIndicator;
+import com.fuzz.indicator.OffSetters;
 import com.fuzz.indicator.text.CutoutTextCell;
 
 /**
