+++ indicator/src/main/java/com/fuzz/indicator/proxy/StateProxy.java
@@ -13,9 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fuzz.indicator;
+package com.fuzz.indicator.proxy;
 
 import android.database.DataSetObserver;
+
+import com.fuzz.indicator.CutoutViewIndicator;
 
 /**
  * Wrap an implementation of this class around your state machine.
