+++ mobile/src/main/java/com/fuzz/emptyhusk/prefab/CVIScrollListener.java
@@ -22,7 +22,7 @@ import android.support.v7.widget.RecyclerView.ViewHolder;
 import android.view.View;
 
 import com.fuzz.indicator.CutoutViewIndicator;
-import com.fuzz.indicator.IndicatorOffsetEvent;
+import com.fuzz.indicator.proxy.IndicatorOffsetEvent;
 
 /**
  * {@link RecyclerView} analog to {@link com.fuzz.indicator.OnViewPagerChangeListener}.
